Working on Envelope::SetNumSamples
Fixed a bug when calculating diff, and continued working on the
interpolation code.

+++ WaveGenerator/Envelope.h
@@ -25,6 +25,6 @@ public:
 
 	void LoadEnvelope(string filename);
 	void SetDuration(double duration);
-	void SetNumSamples(unsigned long numSamples);
+	void SetNumSamples(unsigned int numSamples);
 };
